moved part of enums into dedicated module
