+++ ads/aqua/common/enums.py
@@ -44,21 +44,6 @@ class Tags(str, metaclass=ExtendedEnumMeta):
     BASE_MODEL_CUSTOM = "aqua_custom_base_model"
 
 
-class InferenceContainerType(str, metaclass=ExtendedEnumMeta):
-    CONTAINER_TYPE_VLLM = "vllm"
-    CONTAINER_TYPE_TGI = "tgi"
-
-
-class InferenceContainerTypeKey(str, metaclass=ExtendedEnumMeta):
-    AQUA_VLLM_CONTAINER_KEY = "odsc-vllm-serving"
-    AQUA_TGI_CONTAINER_KEY = "odsc-tgi-serving"
-
-
-class InferenceContainerParamType(str, metaclass=ExtendedEnumMeta):
-    PARAM_TYPE_VLLM = "VLLM_PARAMS"
-    PARAM_TYPE_TGI = "TGI_PARAMS"
-
-
 class RqsAdditionalDetails(str, metaclass=ExtendedEnumMeta):
     METADATA = "metadata"
     CREATED_BY = "createdBy"
@@ -67,24 +52,3 @@ class RqsAdditionalDetails(str, metaclass=ExtendedEnumMeta):
     MODEL_VERSION_SET_NAME = "modelVersionSetName"
     PROJECT_ID = "projectId"
     VERSION_LABEL = "versionLabel"
-
-
-class FineTuningDefinedMetadata(str, metaclass=ExtendedEnumMeta):
-    """Represents the defined metadata keys used in Fine Tuning."""
-
-    VAL_SET_SIZE = "val_set_size"
-    TRAINING_DATA = "training_data"
-
-
-class FineTuningCustomMetadata(str, metaclass=ExtendedEnumMeta):
-    """Represents the custom metadata keys used in Fine Tuning."""
-
-    FT_SOURCE = "fine_tune_source"
-    FT_SOURCE_NAME = "fine_tune_source_name"
-    FT_OUTPUT_PATH = "fine_tune_output_path"
-    FT_JOB_ID = "fine_tune_job_id"
-    FT_JOB_RUN_ID = "fine_tune_jobrun_id"
-    TRAINING_METRICS_FINAL = "train_metrics_final"
-    VALIDATION_METRICS_FINAL = "val_metrics_final"
-    TRAINING_METRICS_EPOCH = "train_metrics_epoch"
-    VALIDATION_METRICS_EPOCH = "val_metrics_epoch"
