+++ ads/aqua/modeldeployment/deployment.py
@@ -9,12 +9,7 @@ from typing import Dict, List
 from oci.data_science.models import ModelDeployment
 
 from ads.aqua.app import AquaApp, logger
-from ads.aqua.common.enums import (
-    InferenceContainerParamType,
-    InferenceContainerType,
-    InferenceContainerTypeKey,
-    Tags,
-)
+from ads.aqua.common.enums import Tags
 from ads.aqua.common.errors import AquaRuntimeError, AquaValueError
 from ads.aqua.common.utils import (
     AQUA_MODEL_TYPE_CUSTOM,
@@ -36,6 +31,11 @@ from ads.aqua.modeldeployment.entities import (
     AquaDeployment,
     AquaDeploymentDetail,
     ContainerSpec,
+)
+from ads.aqua.modeldeployment.enums import (
+    InferenceContainerParamType,
+    InferenceContainerType,
+    InferenceContainerTypeKey,
 )
 from ads.common.object_storage_details import ObjectStorageDetails
 from ads.common.utils import get_log_links
